Use secure defaults for TLS (#78)
* Use secure defaults for TLS

* Make hostname verification work for upstream servers

* Fix a few missed host headers for hostname verification

It is safer to default to secure TLS defaults. Then, users can override the defaults, if necessary.

In the process of testing this PR, I discovered that the proxying of requests failed with the following error:

> [HPM] Error occurred while proxying request pulsar-adminconsole/admin/v2/namespaces/public to https://pulsar-proxy.test.svc.cluster.local:8443/ [ERR_TLS_CERT_ALTNAME_INVALID] (https://nodejs.org/api/errors.
html#errors_common_system_errors)

The certificate had the correct alt name, and after digging deeper, I can see that the issue is a mismatched `Host` header. Here is a stack overflow link that corroborates this theory: https://stackoverflow.com/questions/43175060/proxying-requests-in-node.

The solution is to use the `changeOrigin` option, which replaces the `Host` header with the correct one.

Note: this was not a security vulnerability. It was the opposite. Hostname verification always failed, even when it should have passed.