fix requests to pulsar proxy
If the admin console is placed behind a proxy such as nginx, some proxy header fields may be added to the
request which cause the pulsar proxy to fail to route the request to the appropriate pulsar services.
This change removes these proxy headers from the request before forwarding on to the pulsar proxy, and
sets the authorization header to the configured pulsar token instead of relying on the token sent by the
original request.

Also logs warnings for failed proxy requests.